% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_area.R
\name{bc_area}
\alias{bc_area}
\title{The size of British Columbia}
\usage{
bc_area(what = "total", units = "km2")
}
\arguments{
\item{what}{Which part of BC? One of \code{'total'} (default), \code{'land'}, or \code{'freshwater'}.}

\item{units}{One of \code{'km2'} (square kilometres; default), \code{'m2'} (square metres),
or \code{'ha'} (hectares)}
}
\value{
The area of B.C. in the desired units (numeric vector).
}
\description{
Total area, Land area only, or Freshwater area only, in \eqn{km^2}, \eqn{m^2}, or \code{ha}.
}
\details{
The sizes are from \href{http://www.statcan.gc.ca/tables-tableaux/sum-som/l01/cst01/phys01-eng.htm}{Statistics Canada}
}
\examples{
bc_area("land", "ha")
}

