% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bc_bbox}
\alias{bc_bbox}
\title{Get an extent/bounding box for British Columbia}
\usage{
bc_bbox(class = c("sf", "sp", "raster"))
}
\arguments{
\item{class}{\code{"sf"}, \code{"sp"}, or \code{"raster"}}
}
\value{
an object denoting a bouding box of British Columbia,
of the corresponding class specified in \code{class}
}
\description{
Get an extent/bounding box for British Columbia
}
\examples{
bc_bbox("sf")
bc_bbox("sp")
bc_bbox("raster")
}
