% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_get_data.R
\name{bc_bound}
\alias{bc_bound}
\title{BC Boundary}
\usage{
bc_bound(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{bc_bound} in the desired class
}
\description{
BC Boundary
}
\examples{
\dontrun{
my_layer <- bc_bound()
my_layer_sp <- bc_bound(class = 'sp')
}

}
