% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_hyd_zn.R
\docType{data}
\name{bc_hyd_zn}
\alias{bc_hyd_zn}
\title{Hydrologic Zone Boundaries of British Columbia}
\format{Zones that represent areas of homogeneous hydrologic and geomorphological characteristics}
\source{
Original data from
  \href{https://catalogue.data.gov.bc.ca/dataset/hydrology-hydrologic-zone-boundaries-of-british-columbia}{DataBC},
   under the
  \href{http://www.data.gov.bc.ca/local/dbc/docs/license/OGL-vbc2.0.pdf}{Open
  Government License - British Columbia}.
}
\usage{
bc_hyd_zn
}
\description{
Hydrologic zone boundaries were digitized from 1:600,000 hardcopy maps.
Coastline & islands were copied from 1:50K BC WSA. Linework updated January 2003.
Polygons & Regions built September 20, 2004. Boundaries accuracy is +/- 1 kilometre based
on hydrologic maps, graphical plots, & previous definitions.
}
\keyword{datasets}
