% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_bound.R
\name{bc_neighbours}
\alias{bc_neighbours}
\title{Boundary of British Columbia, provinces/states and the portion of the Pacific Ocean that borders British Columbia}
\source{
\code{bcdata::bcdc_get_data('b9bd93e1-0226-4351-b943-05c6f80bd5da')}
}
\usage{
bc_neighbours(class = deprecated(), ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}.
The \code{class} argument
is deprecated as of bcmaps 1.2.0 and will be removed in Summer 2023. At that point this function will only return an \code{sf} object.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{bc_neighbours} in the desired class
}
\description{
Boundary of British Columbia, provinces/states and the portion of the Pacific Ocean that borders British Columbia
}
\examples{
\dontrun{
my_layer <- bc_neighbours()
}

}
