% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bcmaps-defunct}
\alias{bcmaps-defunct}
\alias{fix_self_intersect}
\title{Defunct functions in bcmaps}
\usage{
fix_self_intersect(...)
}
\description{
These functions are gone, no longer available.
}
\details{
\itemize{
 \item \code{\link{fix_self_intersect}}: This function is defunct. Use \code{\link{fix_geo_problems}} instead.
}
}

