% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmaps-package.R
\docType{package}
\name{bcmaps-package}
\alias{bcmaps}
\alias{bcmaps-package}
\title{bcmaps: Map Layers and Spatial Utilities for British Columbia}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides access to various spatial layers for B.C., such as 
    administrative boundaries, natural resource management boundaries, etc. 
    All layers are imported from the 'bcmaps.rdata' package as 'sf' or 'Spatial' objects
    through function calls in this package. All layers are in B.C. 'Albers' equal-area projection 
    <http://spatialreference.org/ref/epsg/nad83-bc-albers/>, which is the B.C. 
    government standard.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/bcmaps}
  \item Report bugs at \url{https://github.com/bcgov/bcmaps/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gov.bc.ca}

Authors:
\itemize{
  \item Stephanie Hazlitt \email{stephanie.hazlitt@gov.bc.ca}
  \item Sam Albers \email{sam.albers@gov.bc.ca}
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
