% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{bec}
\alias{bec}
\title{British Columbia BEC Map}
\usage{
bec(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{bec} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
See \link[bcmaps.rdata:bec]{bcmaps.rdata::bec} for details.
}
\examples{
\dontrun{
my_layer <- bec()
my_layer_sp <- bec(class = 'sp')
}

}
