% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bec_colours}
\alias{bec_colours}
\alias{bec_colors}
\title{Biogeoclimatic Zone Colours}
\usage{
bec_colours()

bec_colors()
}
\value{
named vector of hexadecimal colour codes. Names are standard
abbreviations of Zone names.
}
\description{
Standard colours used to represent Biogeoclimatic Zone colours
}
