% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cded.R
\name{cded_raster}
\alias{cded_raster}
\title{Get Canadian Digital Elevation Model (CDED) as a \code{raster} object}
\usage{
cded_raster(aoi = NULL, mapsheets = NULL, .predicate = sf::st_intersects, ...)
}
\arguments{
\item{aoi}{Area of Interest. An sf polygon.}

\item{mapsheets}{Mapsheets grid to retrieve raster tiles. Defaults to mapsheets_250K}

\item{.predicate}{geometry predicate function used to find the mapsheets from your aoi. Default \link[sf:geos_binary_pred]{sf::st_intersects}.}

\item{...}{Further arguments passed on to \link[raster:raster]{raster::raster}}
}
\value{
a \code{raster} object of the cded tiles for the specified area of interest
}
\description{
Get Canadian Digital Elevation Model (CDED) as a \code{raster} object
}
\examples{
vic <- census_subdivision()[census_subdivision()$CENSUS_SUBDIVISION_NAME == "Victoria", ]
vic_cded <- cded_raster(aoi = vic)
}
