% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{census_division}
\alias{census_division}
\title{Current Census Division Boundaries}
\source{
\code{bcdata::bcdc_get_data(record = 'ef17918a-597a-4012-8534-f8e71d8735b3', resource = '36b530c2-1de6-44a2-a6f6-c1fce36c53ed')}
}
\usage{
census_division(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{census_division} in the desired class
}
\description{
Current Census Division Boundaries
}
\examples{
\dontrun{
my_layer <- census_division()
my_layer_sp <- census_division(class = 'sp')
}

}
