% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{census_metropolitan_area}
\alias{census_metropolitan_area}
\title{Current Census Metropolitan Areas}
\source{
\code{bcdata::bcdc_get_data(record = 'a6fb34b7-0937-4718-8f1f-43dba2c0f407', resource = 'f129a965-363e-4d7e-8319-a4010e002fdb')}
}
\usage{
census_metropolitan_area(
  class = deprecated(),
  ask = interactive(),
  force = FALSE
)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}.
The \code{class} argument
is deprecated as of bcmaps 1.2.0 and will be removed in Summer 2023. At that point this function will only return an \code{sf} object.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{census_metropolitan_area} in the desired class
}
\description{
Current Census Metropolitan Areas
}
\examples{
\dontrun{
my_layer <- census_metropolitan_area()
}

}
