% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{census_tract}
\alias{census_tract}
\title{Current Census Tract Boundaries}
\source{
\code{bcdata::bcdc_get_data(record = '539aae5b-12f6-4934-9592-9b27acc827f8', resource = 'be767db6-0d4e-4906-a89b-564637f14db9')}
}
\usage{
census_tract(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{census_tract} in the desired class
}
\description{
Current Census Tract Boundaries
}
\examples{
\dontrun{
my_layer <- census_tract()
my_layer_sp <- census_tract(class = 'sp')
}

}
