% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_nr_rd}
\alias{combine_nr_rd}
\title{Combine Northern Rockies Regional Municipality with Regional Districts}
\usage{
combine_nr_rd(class = deprecated())
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. \code{class = sp} is no
longer supported and as of bcmaps 1.2.0 (planned for Summer 2023) the class argument will be removed.}
}
\value{
A layer where the Northern Rockies Regional Municipality has been
combined with the Regional Districts to form a full provincial coverage.
}
\description{
Combine Northern Rockies Regional Municipality with Regional Districts
}
