% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community-boundaries.R
\name{community_councils}
\alias{community_councils}
\title{Community Councils Boundaries}
\usage{
community_councils(ask = interactive(), force = FALSE)
}
\arguments{
\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{community_councils} in the desired class
}
\description{
Community Councils Boundaries
}
\examples{
\dontrun{
my_layer <- community_councils()
}

}
