% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{data_zone_2011}
\alias{data_zone_2011}
\title{Data Zone Boundaries 2011}
\usage{
data_zone_2011(ask = interactive(), force = FALSE)
}
\arguments{
\item{ask}{Should the function ask the user before downloading the data to a
cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{data_zone_2011} in the desired class
}
\description{
Data Zone Boundaries 2011
}
\examples{
\dontrun{
my_layer <- data_zone_2011()
}

}
