% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-utils.R
\name{delete_cache}
\alias{delete_cache}
\title{Delete cached files}
\usage{
delete_cache(files_to_delete = NULL)
}
\arguments{
\item{files_to_delete}{An optional argument to specify which files should be delete.
Defaults to deleting all files pausing for permission from user. If a subset
of files are specified, the files are immediately deleted.}
}
\value{
A logical of whether the file(s) were successful deleted
}
\description{
Delete cached files
}
\examples{
\dontrun{
## See which files you have
show_cached_files()

## Delete your whole cache
delete_cache()

## Specify which files are deleted
delete_cache(c('regional_districts.rds', 'bc_cities.rds'))
}
}
