% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{ecoprovinces}
\alias{ecoprovinces}
\title{British Columbia Ecoprovinces}
\source{
\code{bcdata::bcdc_get_data(record = '51832f47-efdf-4956-837a-45fc2c9032dd', resource = '811fcedb-1a53-4574-8149-454f4a740682')} for direct access to the B.C. Data Catalogue.
}
\usage{
ecoprovinces(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{ecoprovinces} in the desired class
}
\description{
British Columbia Ecoprovinces
}
\examples{
\dontrun{
my_layer <- ecoprovinces()
my_layer_sp <- ecoprovinces(class = 'sp')
}

}
