% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_geo_problems}
\alias{fix_geo_problems}
\title{Check and fix polygons that self-intersect, and sometimes can fix orphan holes}
\usage{
fix_geo_problems(obj)
}
\arguments{
\item{sp_obj}{The SpatialPolygons* object to check/fix}
}
\value{
The SpatialPolygons* object, repaired if necessary
}
\description{
This uses the common method of buffering by zero, using gBuffer in the rgeos package.
}
\details{
\code{fix_self_intersect} has been removed and will no longer work. Use
\code{fix_geo_problems} instead
}

