% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_self_intersect}
\alias{fix_self_intersect}
\title{Check and fix polygons that self-intersect}
\usage{
fix_self_intersect(sp_obj)
}
\arguments{
\item{sp_obj}{The SpatialPolygons* object to check/fix}
}
\value{
The SpatialPolygons* object, repaired if necessary
}
\description{
This uses the common method of buffering by zero, using gBuffer in the rgeos package.
}

