% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_big_data}
\alias{get_big_data}
\title{Download a large data file}
\usage{
get_big_data(
  what,
  class = c("sf", "sp"),
  release = "latest",
  force = FALSE,
  ask = TRUE
)
}
\arguments{
\item{what}{The name of the object to download}

\item{class}{class of object to import; one of \code{"sf"} (default) or \code{"sp"}.}

\item{release}{Specific version of bcmapsdata to get the desired dataset from. Default \code{"latest"}}

\item{force}{Force downloading and overwriting existing dataset. Default \code{FALSE}}

\item{ask}{Ask whether or not to write to the default data directory for bcmaps. Default \code{TRUE}}
}
\description{
Download a large data file
}
