% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_layer}
\alias{get_layer}
\title{Get a scotmaps spatial layer}
\usage{
get_layer(layer, ask = TRUE, force = FALSE)
}
\arguments{
\item{layer}{the name of the layer. The list of available layers can be
obtained by running \code{available_layers()}}

\item{ask}{Should the function ask the user before downloading the data to a
cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
the layer requested
}
\description{
Layers are assessed directly from the \href{https://www.spatialdata.gov.scot/}{spatialdata.gov.scot}
catalogue.
}
\examples{
\dontrun{
get_layer("local_authorities")
}
}
