% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{get_poly_attribute}
\alias{get_poly_attribute}
\title{Get or calculate the attribute of a list-column containing nested dataframes.}
\usage{
get_poly_attribute(x, col, fun, ...)
}
\arguments{
\item{x}{the list-column in the (SpatialPolygons)DataFrame that contains nested data.frames}

\item{col}{the column in the nested data frames from which to retrieve/calculate attributes}

\item{fun}{function to determine the resulting single attribute from overlapping polygons}

\item{...}{other parameters passed on to \code{fun}}
}
\value{
An atomic vector of the same length as x
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function is defunct as of bcmaps 2.0.0.
}
\keyword{internal}
