% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_aquifers.R
\docType{data}
\name{gw_aquifers}
\alias{gw_aquifers}
\title{Polygons delineating British Columbia's developed ground water aquifers (that have been mapped). Most aquifer boundaries
are delineated based on geology, hydrology and topographic information. Some aquifer boundaries stop at the border of BC
mapsheet boundaries due to resource or data constraints at the time of mapping.}
\format{A spatialPolygonsDataFrame with polygons delineating mapped, developed ground water aquifers for British Columbia.}
\source{
Original data from
  \href{https://catalogue.data.gov.bc.ca/dataset/ground-water-aquifers}{},
   under the
  \href{http://www2.gov.bc.ca/gov/content/data/open-data/open-government-license-bc}{Open
  Government License - British Columbia}.
}
\usage{
gw_aquifers
}
\description{
Polygons delineating British Columbia's developed ground water aquifers (that have been mapped). Most aquifer boundaries
are delineated based on geology, hydrology and topographic information. Some aquifer boundaries stop at the border of BC
mapsheet boundaries due to resource or data constraints at the time of mapping.
}
\keyword{datasets}
