% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{health_ha}
\alias{health_ha}
\title{Health Authority Boundaries}
\source{
\code{bcdata::bcdc_get_data(record = '7bc6018f-bb4f-4e5d-845e-c529e3d1ac3b', resource = '93b79a3c-2da4-4fd4-b953-2f5c690db430')}
}
\usage{
health_ha(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{health_ha} in the desired class
}
\description{
Health Authority Boundaries
}
\examples{
\dontrun{
my_layer <- health_ha()
my_layer_sp <- health_ha(class = 'sp')
}

}
