% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{health_hsda}
\alias{health_hsda}
\title{Health Service Delivery Area Boundaries}
\source{
\code{bcdata::bcdc_get_data(record = '71c930b9-563a-46da-a10f-ead49ccbc390', resource = 'c5dad467-229b-4378-852b-ff92479a65b6')}
}
\usage{
health_hsda(ask = interactive(), force = FALSE)
}
\arguments{
\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{health_hsda} as an \code{sf} object.
}
\description{
Health Service Delivery Area Boundaries
}
\examples{
\dontrun{
my_layer <- health_hsda()
}

}
