% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{health_lha}
\alias{health_lha}
\title{Local Health Area Boundaries}
\source{
\code{bcdata::bcdc_get_data(record = 'afd021d9-7722-4410-b506-d394c66e74fc', resource = 'd6e951d3-5103-475a-8bb6-b4d275e6343f')}
}
\usage{
health_lha(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{health_lha} in the desired class
}
\description{
Local Health Area Boundaries
}
\examples{
\dontrun{
my_layer <- health_lha()
my_layer_sp <- health_lha(class = 'sp')
}

}
