% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hydrozones}
\alias{hydrozones}
\title{Hydrologic Zone Boundaries of British Columbia}
\source{
\code{bcdata::bcdc_get_data(record = '329fd234-8835-4d44-9aaa-97c37bfc8d92', resource = 'baeb665e-85c7-4a7b-8e67-8b956785490a')} for direct access to the B.C. Data Catalogue.
}
\usage{
hydrozones(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{hydrozones} in the desired class
}
\description{
Hydrologic Zone Boundaries of British Columbia
}
\examples{
\dontrun{
my_layer <- hydrozones()
my_layer_sp <- hydrozones(class = 'sp')
}

}
