% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{local_authorities}
\alias{local_authorities}
\title{Scottish Local Authorities Boundaries}
\usage{
local_authorities(ask = interactive(), force = FALSE)
}
\arguments{
\item{ask}{Should the function ask the user before downloading the data to a
cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{local_authorities} in the desired class
}
\description{
Scottish Local Authorities Boundaries
}
\examples{
\dontrun{
my_layer <- local_authorities()
}

}
