% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{mapsheets_250K}
\alias{mapsheets_250K}
\title{NTS 250K Grid - Digital Baseline Mapping at 1:250,000 (NTS)}
\source{
\code{bcdata::bcdc_get_data(record = 'c50803db-7645-4da8-bce5-1bc5b501a7b3', resource = '3a8bcd6d-2414-412f-ad9f-59c69d7ed165')}
}
\usage{
mapsheets_250K(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{mapsheets_250K} in the desired class
}
\description{
NTS 250K Grid - Digital Baseline Mapping at 1:250,000 (NTS)
}
\examples{
\dontrun{
my_layer <- mapsheets_250K()
my_layer_sp <- mapsheets_250K(class = 'sp')
}

}
