% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-non-catalogue-data.R
\docType{data}
\name{mapsheets_250K_data}
\alias{mapsheets_250K_data}
\title{Raw NTS 250K Grid - Digital Baseline Mapping at 1:250,000 (NTS)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 112 rows and 8 columns.
}
\usage{
mapsheets_250K_data
}
\description{
Internal data.
}
\keyword{datasets}
