% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-non-catalogue-data.R
\name{mapsheets_50K}
\alias{mapsheets_50K}
\title{NTS 50K Grid - Digital Baseline Mapping at 1:50,000 (NTS)}
\source{
https://open.canada.ca/data/en/dataset/055919c2-101e-4329-bfd7-1d0c333c0e62
}
\usage{
mapsheets_50K()
}
\value{
The spatial layer of \code{mapsheets_50K} as an \code{sf} object.
}
\description{
NTS 50K Grid - Digital Baseline Mapping at 1:50,000 (NTS)
}
\examples{
\dontrun{
my_layer <- mapsheets_50K()
}

}
