% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{marine_areas}
\alias{marine_areas}
\title{Scottish Marine Area}
\usage{
marine_areas(ask = interactive(), force = FALSE)
}
\arguments{
\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{marine_areas} in the desired class
}
\description{
Scottish Marine Area
}
\examples{
\dontrun{
my_layer <- marine_areas()
}

}
