% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{munipalities}
\alias{munipalities}
\title{British Columbia Municipalities}
\source{
\code{bcdata::bcdc_get_data(record = 'e3c3c580-996a-4668-8bc5-6aa7c7dc4932', resource = '25c95b07-5882-47ff-970d-36cb243b8355')} for direct access to the B.C. Data Catalogue.
}
\usage{
munipalities(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{munipalities} in the desired class
}
\description{
British Columbia Municipalities
}
\examples{
\dontrun{
my_layer <- munipalities()
my_layer_sp <- munipalities(class = 'sp')
}

}
