% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr_areas.R
\docType{data}
\name{nr_areas}
\alias{nr_areas}
\title{British Columbia Natural Resource (NR) Areas}
\format{A spatialPolygonsDataFrame with the boundaries of British Columbia's
  3 Natural Resource (NR) Areas.}
\source{
Original data from
  \href{https://catalogue.data.gov.bc.ca/dataset/natural-resource-nr-areas}{DataBC},
   under the
  \href{http://www.data.gov.bc.ca/local/dbc/docs/license/OGL-vbc2.0.pdf}{Open
  Government License - British Columbia}.
}
\usage{
nr_areas
}
\description{
British Columbia Natural Resource (NR) Areas are administrative boundaries
established by the Ministry of Forests, Lands and Natural Resource Operations.
These boundaries are an amalgamation of all NR Regions within the supplied boundaries.
}
\keyword{datasets}
