% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr_districts.R
\docType{data}
\name{nr_districts}
\alias{nr_districts}
\title{British Columbia Natural Resource (NR) Districts}
\format{A spatialPolygonsDataFrame with the boundaries of British Columbia's
   Natural Resource (NR) Districts.}
\source{
Original data from
  \href{https://catalogue.data.gov.bc.ca/dataset/natural-resource-nr-district}{DataBC},
   under the
  \href{http://www.data.gov.bc.ca/local/dbc/docs/license/OGL-vbc2.0.pdf}{Open
  Government License - British Columbia}.
}
\usage{
nr_districts
}
\description{
British Columbia Natural Resource (NR) Districts are administrative boundaries
established by the Ministry of Forests, Lands and Natural Resource Operations.
}
\keyword{datasets}

