% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotmaps.R
\docType{package}
\name{scotmaps}
\alias{scotmaps}
\title{scotmaps: A data package providing open licence map layers for Scotland}
\description{
Various layers of Scotland, including administrative boundaries, natural
resource management boundaries, etc. All layers are available as
\href{https://CRAN.R-project.org/package=sf}{sf} objects, and sourced from
\href{https://www.spatialdata.gov.scot/}{spatialdata.gov.scot} in \href{https://spatialreference.org/ref/epsg/27700}{British National Grid 27700}. Each
layer's individual help page contains a section describing the source for the
data. This package is modified from \code{bcmaps} package.
}
