% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{self_union}
\alias{self_union}
\title{Union a SpatialPolygons* object with itself to remove overlaps, while retaining attributes}
\usage{
self_union(x)
}
\arguments{
\item{x}{A \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} object}
}
\value{
A \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated as of bcmaps 1.2.0, and will be removed in Summer 2023.
Use \code{raster::union()} for \code{SpatialPolygonsDataFrame}s, or
\code{sf::st_union()} with \code{sf} objects instead.

The IDs of source polygons are stored in a list-column called
\code{union_ids}, and original attributes (if present) are stored as nested
dataframes in a list-column called \code{union_df}.
}
\keyword{internal}
