% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-utils.R
\name{show_cached_files}
\alias{show_cached_files}
\title{Show path of which files are stored locally}
\usage{
show_cached_files()
}
\description{
Show path of which files are stored locally
}
\examples{
\dontrun{
show_cached_files()
}
}
