% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_by_poly.R
\name{summarize_raster_list}
\alias{summarize_raster_list}
\title{Summarize a list of rasters into a list of numeric vectors}
\usage{
summarize_raster_list(raster_list, parallel = FALSE,
  future_strategy = "multisession", workers = NULL, ...)
}
\arguments{
\item{raster_list}{list of rasters}

\item{parallel}{process in parallel? Default \code{FALSE}. Not currently
available on Windows.}

\item{future_strategy}{the strategy to use in \code{future::plan()} for parallel
computation. Defaults to \code{"multisession"}.}

\item{workers}{number of workers if doing parallel. Default \code{NULL} uses
\code{future::availableCores()}}

\item{...}{passed on to \code{future::plan()}}
}
\value{
a list of numeric vectors
}
\description{
Summarize a list of rasters into a list of numeric vectors
}
