% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_bc_albers}
\alias{transform_bc_albers}
\title{Transform a Spatial* object to BC Albers projection}
\usage{
transform_bc_albers(sp_obj)
}
\arguments{
\item{sp_obj}{The Spatial* object to transform}
}
\value{
the Spatial* object in BC Albers projection
}
\description{
Transform a Spatial* object to BC Albers projection
}

