% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm-convert.R
\name{utm_convert}
\alias{utm_convert}
\title{Convert a data.frame of UTM coordinates to a single CRS}
\usage{
utm_convert(x, xcol, ycol, zonecol, crs = "EPSG:3005")
}
\arguments{
\item{x}{data.frame containing UTM coordinates, with a zone column}

\item{xcol}{the name of the 'easting' columnm}

\item{ycol}{the name of the 'northing' column}

\item{zonecol}{the name of the 'zone' column}

\item{crs}{target CRS. Default BC Albers (EPSG:3005)}
}
\value{
sf object in the chosen CRS
}
\description{
Convert a data.frame of UTM coordinates to a single CRS
}
