% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cded.R
\name{vrt_files}
\alias{vrt_files}
\title{List the files that a vrt is built on}
\usage{
vrt_files(vrt, omit_vrt = FALSE)
}
\arguments{
\item{vrt}{path to a .vrt file}

\item{omit_vrt}{omit the listing of the original vrt. Default \code{FALSE}}
}
\value{
character vector of tiles
}
\description{
List the files that a vrt is built on
}
