% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cded.R
\name{vrt_info}
\alias{vrt_info}
\title{Get metadata about a .vrt file}
\usage{
vrt_info(vrt, options = character(0), quiet = FALSE)
}
\arguments{
\item{vrt}{path to a .vrt file}

\item{options}{options to pass to \code{gdalinfo}. See
\href{https://gdal.org/en/stable/programs/gdalinfo.html}{here} for
possible options.}

\item{quiet}{suppress output to the console (default
\code{FALSE})}
}
\value{
character of vrt metadata
}
\description{
Get metadata about a .vrt file
}
