% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{water_districts}
\alias{water_districts}
\title{Polygons delineating British Columbia's Water Management Districts. Water Districts are administrative areas established by the WATER REGULATIONS.}
\usage{
water_districts(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{water_districts} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
See \link[bcmaps.rdata:water_districts]{bcmaps.rdata::water_districts} for details.
}
\examples{
\dontrun{
my_layer <- water_districts()
my_layer_sp <- water_districts(class = 'sp')
}

}
