% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{water_monitoring_sites}
\alias{water_monitoring_sites}
\title{Water monitoring sites}
\usage{
water_monitoring_sites(ask = interactive(), force = FALSE)
}
\arguments{
\item{ask}{Should the function ask the user before downloading the data to a
cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{water_monitoring_sites} in the desired class
}
\description{
Layers are assessed directly from the \href{https://map.sepa.org.uk/arcgis/rest/services/}{SEPA}
catalogue.
}
\examples{
\dontrun{
my_layer <- water_monitoring_sites()
}

}
