% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watercourses_15M.R
\docType{data}
\name{watercourses_15M}
\alias{watercourses_15M}
\title{British Columbia watercourses at 1:15M scale.}
\format{A spatialPolygonsDataFrame with watercourses at the 1:15M scale for British Columbia.}
\source{
Original data from
  \href{https://www.nrcan.gc.ca/earth-sciences/geography/topographic-information/free-data-geogratis/download-directory-documentation/17215}{},
   under the
  \href{http://open.canada.ca/en/open-government-licence-canada}{Open
  Government License - Canada}.
}
\usage{
watercourses_15M
}
\description{
Watercourses at 1:15M scale for British Columbia sourced from Natural Resources Canada GeoGratis data.
}
\keyword{datasets}

