% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-non-catalogue-data.R
\name{watercourses_5M}
\alias{watercourses_5M}
\title{British Columbia watercourses at 1:5M scale}
\source{
https://ftp.maps.canada.ca/pub/nrcan_rncan/vector/canvec/fgdb/Hydro/canvec_5M_CA_Hydro_fgdb.zip
}
\usage{
watercourses_5M(ask = interactive(), force = FALSE)
}
\arguments{
\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{watercourses_5M} as an \code{sf} object.
}
\description{
British Columbia watercourses at 1:5M scale
}
\examples{
\dontrun{
my_layer <- watercourses_5M()
}

}
