% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{watersheds}
\alias{watersheds}
\title{British Columbia Hydrometric Watershed Boundaries}
\usage{
watersheds(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{watersheds} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
See \link[bcmaps.rdata:watersheds]{bcmaps.rdata::watersheds} for details.
}
\examples{
\dontrun{
my_layer <- watersheds()
my_layer_sp <- watersheds(class = 'sp')
}

}
