% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airzones.R
\docType{data}
\name{airzones}
\alias{airzones}
\title{British Columbia Air Zones}
\format{A spatialPolygonsDataFrame with the boundaries of British Columbia's
  seven air zones.

  The data slot contains one column, \code{Airzone}, with the name of the Air Zone.}
\source{
Original data from \href{http://catalogue.data.gov.bc.ca/dataset/british-columbia-air-zones}{DataBC},
under the \href{http://www.data.gov.bc.ca/local/dbc/docs/license/OGL-vbc2.0.pdf}{Open Government License - British Columbia}.
}
\usage{
airzones
}
\description{
Boundaries of British Columbia Air Zones, which are areas of the province
for reporting on Canadian Ambient Air Quality Standards
(\href{http://www.ccme.ca/en/current_priorities/air/caaqs.html}{CAAQS}), and
implementing the Air Quality Management System. For more information on Air
Zones in B.C. see \url{http://www.bcairquality.ca/plans/national-air-quality-management-system.html}.
}
\keyword{datasets}

