% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{airzones}
\alias{airzones}
\title{British Columbia Air Zones}
\source{
\code{bcdata::bcdc_get_data(record = 'e8eeefc4-2826-47bc-8430-85703d328516', resource = 'c495d082-b586-4df0-9e06-bd6b66a8acd9')}
}
\usage{
airzones(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{airzones} in the desired class
}
\description{
British Columbia Air Zones
}
\examples{
\dontrun{
my_layer <- airzones()
my_layer_sp <- airzones(class = 'sp')
}

}
