% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{airzones}
\alias{airzones}
\title{British Columbia Air Zones}
\usage{
airzones(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{airzones} in the desired class
}
\description{
You must have the \code{bcmapsdata} package installed to use this function.
}
\details{
Type \code{?bcmapsdata::airzones} for details.
}
\examples{
\dontrun{
my_layer <- airzones()
my_layer_sp <- airzones(class = 'sp')
}

}
