% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{available_layers}
\alias{available_layers}
\title{List available data layers}
\usage{
available_layers()
}
\value{
A data.frame of layers, with Titles, and an additional column \code{shortcut_function}
containing the name of the shortcut function that can be used to return the
layer. A value of \code{NA} in this column means the layer is available via \code{get_data()} but
there is no shortcut function for it.
}
\description{
A data.frame of all available layers in the bcmaps package. This drawn
directly from the bcmaps.rdata package and will therefore be the most current list
layers available.
}
\examples{
\dontrun{
available_layers()
}
}
