% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bc_bbox}
\alias{bc_bbox}
\title{Get an extent/bounding box for British Columbia}
\usage{
bc_bbox(class = c("sf", "raster"), crs = 3005)
}
\arguments{
\item{class}{\code{"sf"}, \code{"raster"}.}

\item{crs}{coordinate reference system: integer with the EPSG code,
or character with proj4string. Default \code{3005} (BC Albers).}
}
\value{
an object denoting a bounding box of British Columbia,
of the corresponding class specified in \code{class}.
}
\description{
Get an extent/bounding box for British Columbia
}
\examples{
\dontrun{
  bc_bbox("sf")
  bc_bbox("raster")
  }
}
