% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_bound.R
\docType{data}
\name{bc_bound}
\alias{bc_bound}
\title{BC Boundary}
\format{A spatialPolygonsDataFrame defining the boundary of each British Columbia. Scale is 1:7.5M

  The data slot contains three columns:
\describe{
  \item{\code{island}}{Name of island, if applicable}
  \item{\code{island_e}}{English name of island, if applicable}
  \item{\code{island_f}}{French name of island, if applicable}
}}
\source{
1:7.5M Atlas of Canada base maps, downloaded from
\href{http://open.canada.ca/data/en/dataset/f77c2027-ed4a-5f6e-9395-067af3e9fc1e}{The Canadian Open Data Portal} on 2015-05-04
under the \href{http://open.canada.ca/en/open-government-licence-canada}{Open Government License - Canada} version 2.0.
}
\usage{
bc_bound
}
\description{
Boundary of British Columbia at 1:7.5M scale
}
\keyword{datasets}

