% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{bc_bound}
\alias{bc_bound}
\title{BC Boundary}
\usage{
bc_bound(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{bc_bound} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
Type \code{?bcmaps.rdata::bc_bound} for details.
}
\examples{
\dontrun{
my_layer <- bc_bound()
my_layer_sp <- bc_bound(class = 'sp')
}

}
