% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_bound.R
\name{bc_bound}
\alias{bc_bound}
\title{BC Boundary}
\source{
\code{bcdata::bcdc_get_data('b9bd93e1-0226-4351-b943-05c6f80bd5da')}
}
\usage{
bc_bound(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{bc_bound} in the desired class
}
\description{
BC Boundary
}
\examples{
\dontrun{
my_layer <- bc_bound()
my_layer_sp <- bc_bound(class = 'sp')
}

}
