% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_bound_hres.R
\docType{data}
\name{bc_bound_hres}
\alias{bc_bound_hres}
\title{BC Boundary}
\format{A spatialPolygonsDataFrame defining the boundary of each British Columbia. Scale is 1:7.5M}
\source{
Adapted from Statistics Canada, Provinces/Territories Boundary File 2011 Census (gpr_000b11a_e), 2016-01-15.
This does not constitute an endorsement by Statistics Canada of this product.
  \url{http://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/files-fichiers/gpr_000b11a_e.zip}
  Available under the \href{http://www.statcan.gc.ca/eng/reference/licence-eng}{Statistics Canada Open License Agreement}
}
\usage{
bc_bound_hres
}
\description{
Boundary of British Columbia
}
\keyword{datasets}

