% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{bc_bound_hres}
\alias{bc_bound_hres}
\title{BC Boundary - High Resolution}
\usage{
bc_bound_hres(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{bc_bound_hres} in the desired class
}
\description{
You must have the \code{bcmapsdata} package installed to use this function.
}
\details{
Type \code{?bcmapsdata::bc_bound_hres} for details.
}
\examples{
\dontrun{
my_layer <- bc_bound_hres()
my_layer_sp <- bc_bound_hres(class = 'sp')
}

}
