% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{bc_cities}
\alias{bc_cities}
\title{BC Major Cities Points 1:2,000,000 (Digital Baseline Mapping)}
\usage{
bc_cities(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{bc_cities} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
Type \code{?bcmaps.rdata::bc_cities} for details.
}
\examples{
\dontrun{
my_layer <- bc_cities()
my_layer_sp <- bc_cities(class = 'sp')
}

}
