% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{bc_neighbours}
\alias{bc_neighbours}
\title{Boundary of British Columbia, provinces/states and the portion of the Pacific Ocean that borders British Columbia}
\usage{
bc_neighbours(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{bc_neighbours} in the desired class
}
\description{
You must have the \code{bcmapsdata} package installed to use this function.
}
\details{
Type \code{?bcmapsdata::bc_neighbours} for details.
}
\examples{
\dontrun{
my_layer <- bc_neighbours()
my_layer_sp <- bc_neighbours(class = 'sp')
}

}
