% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmaps-package.R
\docType{package}
\name{bcmaps-package}
\alias{bcmaps}
\alias{bcmaps-package}
\title{bcmaps: Map Layers and Spatial Utilities for British Columbia}
\description{
Provides access to various spatial layers for B.C., such as administrative boundaries, natural resource management boundaries, etc. Most layers are imported from the 'bcdata' package as 'sf' or 'Spatial' objects through function calls in this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/bcmaps}
  \item Report bugs at \url{https://github.com/bcgov/bcmaps/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})

Authors:
\itemize{
  \item Sam Albers \email{sam.albers@gmail.com} (\href{https://orcid.org/0000-0002-9270-7884}{ORCID}) [contributor]
  \item Stephanie Hazlitt \email{stephhazlitt@gmail.com} (\href{https://orcid.org/0000-0002-3161-2304}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
