% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmaps-package.R
\docType{package}
\name{bcmaps-package}
\alias{bcmaps}
\alias{bcmaps-package}
\title{bcmaps: Map Layers and Spatial Utilities for British Columbia}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Various layers of B.C., including administrative boundaries, natural resource management boundaries, census boundaries etc. All layers are available in BC Albers (\url{https://spatialreference.org/ref/epsg/nad83-bc-albers/}) equal-area projection, which is the B.C. government standard. The layers are sourced from the British Columbia and Canadian government under open licenses, including B.C. Data Catalogue (\url{https://data.gov.bc.ca}), the Government of Canada Open Data Portal (\url{https://open.canada.ca/en/using-open-data}), and Statistics Canada (\url{https://www.statcan.gc.ca/en/reference/licence}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/bcmaps}
  \item \url{https://bcgov.github.io/bcmaps/}
  \item Report bugs at \url{https://github.com/bcgov/bcmaps/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})

Authors:
\itemize{
  \item Sam Albers \email{sam.albers@gmail.com} (\href{https://orcid.org/0000-0002-9270-7884}{ORCID}) [contributor]
  \item Stephanie Hazlitt \email{stephhazlitt@gmail.com} (\href{https://orcid.org/0000-0002-3161-2304}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
