% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmaps.R
\docType{package}
\name{bcmaps}
\alias{bcmaps}
\title{bcmaps: A data package providing various map layers for British Columbia}
\description{
Various layers of B.C., including administrative boundaries, natural
resource management boundaries, etc. All layers are available as both \href{https://CRAN.R-project.org/package=sf}{sf}
and \href{https://CRAN.R-project.org/package=sp}{Spatial}  objects, and are in
\href{https://spatialreference.org/ref/epsg/nad83-bc-albers/}{BC Albers}
equal-area projection, which is the B.C. government standard.
The layers are sourced from the British Columbia and Canadian government under open licenses, including
\href{https://data.gov.bc.ca}{DataBC}, the Government of Canada
\href{https://open.canada.ca/en/open-data}{Open Data Portal}, and
\href{https://www.statcan.gc.ca/eng/reference/licence}{Statistics Canada}.
Each layer's individual help page contains a section describing the source for the data.
}
