% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmaps.R
\docType{package}
\name{bcmaps}
\alias{bcmaps}
\alias{bcmaps-package}
\title{bcmaps: A data package providing various map layers for British Columbia}
\description{
Various layers of B.C., including administrative boundaries, natural
resource management boundaries, etc. All layers are available as both \href{https://CRAN.R-project.org/package=sf}{sf}
and \href{https://CRAN.R-project.org/package=sp}{Spatial}  objects, and are in
\href{http://spatialreference.org/ref/epsg/nad83-bc-albers/}{BC Albers}
projection, which is the
\href{https://www.for.gov.bc.ca/hts/risc/pubs/other/mappro/index.htm}{B.C. government standard}.
The layers are sourced from the British Columbia and Canadian government under open licenses, including
\href{http://data.gov.bc.ca}{DataBC}, the Government of Canada
\href{http://open.canada.ca/en/open-data}{Open Data Portal}, and
\href{http://www.statcan.gc.ca/eng/reference/licence-eng}{Statistics Canada}.
Each layer's individual help page contains a section describing the source for the data.
}
