% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{bec}
\alias{bec}
\title{British Columbia BEC Map}
\format{
An \code{sf} or \code{Spatial} polygons object with B.C.'s Biogeoclimatic Ecosystem
Classification (BEC) Zone/Subzone/Variant/Phase map
}
\source{
Original data from the
\href{https://catalogue.data.gov.bc.ca/dataset/f358a53b-ffde-4830-a325-a5a03ff672c3}{B.C. Data Catalogue},
under the
\href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open
Government Licence - British Columbia}.
}
\usage{
bec(class = c("sf", "sp"), ...)
}
\arguments{
\item{class}{class of object to import; one of \code{"sf"} (default) or \code{"sp"}.}

\item{...}{arguments passed on to \link{get_big_data}}
}
\description{
The current and most detailed version of the approved corporate provincial
digital Biogeoclimatic Ecosystem Classification (BEC)
Zone/Subzone/Variant/Phase map (version 10, August 31st, 2016).
}
