% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{census_dissemination_area}
\alias{census_dissemination_area}
\title{Current Census Dissemination Areas}
\source{
\code{bcdata::bcdc_get_data(record = 'a091fd65-d682-4a24-8c0e-68de7c87e3a3', resource = 'a7fa66d4-0f95-4c58-861d-42d875e5b79f')}
}
\usage{
census_dissemination_area(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{census_dissemination_area} in the desired class
}
\description{
Current Census Dissemination Areas
}
\examples{
\dontrun{
my_layer <- census_dissemination_area()
my_layer_sp <- census_dissemination_area(class = 'sp')
}

}
