% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{census_economic}
\alias{census_economic}
\title{Current Census Economic Region Boundaries}
\source{
\code{bcdata::bcdc_get_data(record = '1aebc451-a41c-496f-8b18-6f414cde93b7', resource = '3f0236cf-b1a1-4f1a-8e9e-86c2c3daff96')}
}
\usage{
census_economic(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{census_economic} in the desired class
}
\description{
Current Census Economic Region Boundaries
}
\examples{
\dontrun{
my_layer <- census_economic()
my_layer_sp <- census_economic(class = 'sp')
}

}
