% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{census_subdivision}
\alias{census_subdivision}
\title{Current Census Subdivision Boundaries}
\source{
\code{bcdata::bcdc_get_data(record = '4c5618c6-38dd-4a62-a3de-9408b4974bb6', resource = '98bd1222-57bb-4504-92c2-4a5857ae7671')}
}
\usage{
census_subdivision(class = deprecated(), ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. \code{class = sp} is no
longer supported and as of bcmaps 1.2.0 (planned for Summer 2023) the class argument will be removed.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{census_subdivision} in the desired class
}
\description{
Current Census Subdivision Boundaries
}
\examples{
\dontrun{
my_layer <- census_subdivision()
}

}
