% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_nr_rd}
\alias{combine_nr_rd}
\title{Combine Northern Rockies Regional Municipality with Regional Districts}
\usage{
combine_nr_rd(class = c("sf", "sp"))
}
\arguments{
\item{class}{The class of the layer returned. Can be either \code{"sf"} (default) or \code{"sp"}}
}
\value{
A layer where the Northern Rockies Regional Municipality has been
combined with the Regional Districts to form a full provincial coverage.
}
\description{
Combine Northern Rockies Regional Municipality with Regional Districts
}
