% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoprovinces.R
\docType{data}
\name{ecoprovinces}
\alias{ecoprovinces}
\title{British Columbia Ecoprovinces}
\format{A spatialPolygonsDataFrame with the boundaries of British Columbia's
  ten ecoprovinces.}
\source{
Original data from
  \href{http://catalogue.data.gov.bc.ca/dataset/ecoprovinces-ecoregion-ecosystem-classification-of-british-columbia}{DataBC},
   under the
  \href{http://www.data.gov.bc.ca/local/dbc/docs/license/OGL-vbc2.0.pdf}{Open
  Government License - British Columbia}.
}
\usage{
ecoprovinces
}
\description{
An ecoprovince is an area with consistent climatic or oceanography, relief
and regional landforms. There are one oceanic, two marine / terrestrial and seven
terrestrial ecoprovinces occurring within British Columbia. Ecoprovinces are
meant to be mapped at 1:2,000,000 for use in provincial state of the
environment reporting
}
\keyword{datasets}

