% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{ecoprovinces}
\alias{ecoprovinces}
\title{British Columbia Ecoprovinces}
\source{
\code{bcdata::bcdc_get_data(record = '51832f47-efdf-4956-837a-45fc2c9032dd', resource = '811fcedb-1a53-4574-8149-454f4a740682')}
}
\usage{
ecoprovinces(class = deprecated(), ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}.
The \code{class} argument
is deprecated as of bcmaps 1.2.0 and will be removed in Summer 2023. At that point this function will only return an \code{sf} object.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{ecoprovinces} in the desired class
}
\description{
British Columbia Ecoprovinces
}
\examples{
\dontrun{
my_layer <- ecoprovinces()
}

}
