% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{ecoprovinces}
\alias{ecoprovinces}
\title{British Columbia Ecoprovinces}
\usage{
ecoprovinces(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{ecoprovinces} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
Type \code{?bcmaps.rdata::ecoprovinces} for details.
}
\examples{
\dontrun{
my_layer <- ecoprovinces()
my_layer_sp <- ecoprovinces(class = 'sp')
}

}
