% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions.R
\docType{data}
\name{ecoregions}
\alias{ecoregions}
\title{British Columbia Ecoregions}
\format{A spatialPolygonsDataFrame with the boundaries of British Columbia's
  43 ecoregions.}
\source{
Original data from
  \href{http://catalogue.data.gov.bc.ca/dataset/ecoregions-ecoregion-ecosystem-classification-of-british-columbia}{DataBC},
   under the
  \href{http://www.data.gov.bc.ca/local/dbc/docs/license/OGL-vbc2.0.pdf}{Open
  Government License - British Columbia}.
}
\usage{
ecoregions
}
\description{
An ecoregion is an area with major physiographic and minor macroclimatic or
oceanographic variation. There are 43 ecoregion in British Columbia of which
39 are terrestrial. Ecodivsions are meant to be mapped at 1:500,000 for
regional strategic planning
}
\keyword{datasets}

