% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{ecosections}
\alias{ecosections}
\title{British Columbia Ecosections}
\usage{
ecosections(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{ecosections} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
Type \code{?bcmaps.rdata::ecosections} for details.
}
\examples{
\dontrun{
my_layer <- ecosections()
my_layer_sp <- ecosections(class = 'sp')
}

}
