% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_geo_problems}
\alias{fix_geo_problems}
\title{Check and fix polygons that self-intersect, and sometimes can fix orphan holes}
\usage{
fix_geo_problems(obj, tries = 5)
}
\arguments{
\item{obj}{The SpatialPolygons* or sf object to check/fix.}

\item{tries}{The maximum number of attempts to repair the geometry. Ignored for \code{sf} objects.}
}
\value{
The \verb{SpatialPolygons*} or \code{sf} object, repaired if necessary
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}.

This function is deprecated as of bcmaps 1.2.0 because it relies on \code{rgeos}
for operations on \code{Spatial} objects, which is being retired. It will be removed
completely in Summer 2023.
For \code{sf} objects simply use \code{sf::st_make_valid()}
}
\keyword{internal}
