% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_geo_problems}
\alias{fix_geo_problems}
\title{Check and fix polygons that self-intersect, and sometimes can fix orphan holes}
\usage{
fix_geo_problems(obj, tries = 5)
}
\arguments{
\item{obj}{The SpatialPolygons* or sf object to check/fix}

\item{tries}{The maximum number of attempts to repair the geometry. Ignored for \code{sf} objects.}
}
\value{
The \verb{SpatialPolygons*} or \code{sf} object, repaired if necessary
}
\description{
For \code{sf} objects, uses \code{sf::st_make_valid}.
Otherwise, uses the common method of buffering by zero.
}
\details{
\code{fix_self_intersect} has been removed and will no longer work. Use
\code{fix_geo_problems} instead
}
