% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{fix_geo_problems}
\alias{fix_geo_problems}
\title{Check and fix polygons that self-intersect, and sometimes can fix orphan holes}
\usage{
fix_geo_problems(obj)
}
\arguments{
\item{obj}{The SpatialPolygons* or sf object to check/fix.}
}
\value{
The \verb{SpatialPolygons*} or \code{sf} object, repaired if necessary
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}.

This function is defunct as of bcmaps 2.0.0.
For \code{sf} objects simply use \code{sf::st_make_valid()}
}
\keyword{internal}
