% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_layer}
\alias{get_layer}
\title{Get a B.C. spatial layer}
\usage{
get_layer(layer, class = deprecated(), ask = interactive(), force = FALSE)
}
\arguments{
\item{layer}{the name of the layer. The list of available layers can be
obtained by running \code{available_layers()}}

\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}.
The \code{class} argument
is deprecated as of bcmaps 1.2.0 and will be removed in Summer 2023. At that point this function will only return an \code{sf} object.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
the layer requested
}
\description{
Get a B.C. spatial layer
}
\examples{
\dontrun{
 get_layer("bc_bound_hres")
}
}
