% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_layer}
\alias{get_layer}
\title{Get a B.C. spatial layer}
\usage{
get_layer(layer, class = c("sf", "sp"), ...)
}
\arguments{
\item{layer}{the name of the layer. The list of available layers can be
obtained by running \code{available_layers()}}

\item{class}{The class of the layer returned. Can be either \code{"sf"} (default) or \code{"sp"}}

\item{...}{arguments passed on to \link{get_big_data} if the layer needs to be downloaded. Ignored if the
layer is available locally in \code{bcmapsdata}.}
}
\value{
the layer requested
}
\description{
Get a B.C. spatial layer
}
\examples{
\dontrun{
 get_layer("bc_bound_hres")

 # As a "Spatial" (sp) object
 get_layer("watercourses_15M")
}
}
