% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_poly_attribute}
\alias{get_poly_attribute}
\title{Get or calculate the attribute of a list-column containing nested dataframes.}
\usage{
get_poly_attribute(x, col, fun, ...)
}
\arguments{
\item{x}{the list-column in the (SpatialPolygons)DataFrame that contains nested data.frames}

\item{col}{the column in the nested data frames from which to retrieve/calculate attributes}

\item{fun}{function to determine the resulting single attribute from overlapping polygons}

\item{...}{other parameters passed on to \code{fun}}
}
\value{
An atomic vector of the same length as x
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated as of bcmaps 1.2.0 because it had a very niche application for
calculating attributes on a \code{SpatialPolygonsDataFrame}, which we are removing
support for. It will be removed completely in Summer 2023.

For example, \code{self_union} produces a \code{SpatialPolygonsDataFrame}
that has a column called \code{union_df}, which contains a \code{data.frame}
for each polygon with the attributes from the constituent polygons.
}
\keyword{internal}
