% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{gw_aquifers}
\alias{gw_aquifers}
\title{British Columbia's developed ground water aquifers}
\source{
\code{bcdata::bcdc_get_data(record = '099d69c5-1401-484d-9e19-c121ccb7977c', resource = '8f421e3a-ccd3-4fab-8198-53ad6e9e2af2')}
}
\usage{
gw_aquifers(class = deprecated(), ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}.
The \code{class} argument
is deprecated as of bcmaps 1.2.0 and will be removed in Summer 2023. At that point this function will only return an \code{sf} object.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{gw_aquifers} in the desired class
}
\description{
British Columbia's developed ground water aquifers
}
\examples{
\dontrun{
my_layer <- gw_aquifers()
}

}
