% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{gw_aquifers}
\alias{gw_aquifers}
\title{British Columbia's developed ground water aquifers}
\usage{
gw_aquifers(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{gw_aquifers} in the desired class
}
\description{
You must have the \code{bcmapsdata} package installed to use this function.
}
\details{
Type \code{?bcmapsdata::gw_aquifers} for details.
}
\examples{
\dontrun{
my_layer <- gw_aquifers()
my_layer_sp <- gw_aquifers(class = 'sp')
}

}
