% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{health_chsa}
\alias{health_chsa}
\title{Community Health Service Areas - CHSA}
\source{
\code{bcdata::bcdc_get_data(record = '68f2f577-28a7-46b4-bca9-7e9770f2f357', resource = '59065b51-511a-4976-b77f-034168365273')}
}
\usage{
health_chsa(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{health_chsa} in the desired class
}
\description{
Community Health Service Areas - CHSA
}
\examples{
\dontrun{
my_layer <- health_chsa()
my_layer_sp <- health_chsa(class = 'sp')
}

}
