% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{health_hsda}
\alias{health_hsda}
\title{Health Service Delivery Area Boundaries}
\source{
\code{bcdata::bcdc_get_data(record = '71c930b9-563a-46da-a10f-ead49ccbc390', resource = 'c5dad467-229b-4378-852b-ff92479a65b6')}
}
\usage{
health_hsda(class = deprecated(), ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}.
The \code{class} argument
is deprecated as of bcmaps 1.2.0 and will be removed in Summer 2023. At that point this function will only return an \code{sf} object.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{health_hsda} in the desired class
}
\description{
Health Service Delivery Area Boundaries
}
\examples{
\dontrun{
my_layer <- health_hsda()
}

}
