% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hydrozones}
\alias{hydrozones}
\title{Hydrologic Zone Boundaries of British Columbia}
\usage{
hydrozones(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{hydrozones} in the desired class
}
\description{
You must have the \code{bcmapsdata} package installed to use this function.
}
\details{
Type \code{?bcmapsdata::hydrozones} for details.
}
\examples{
\dontrun{
my_layer <- hydrozones()
my_layer_sp <- hydrozones(class = 'sp')
}

}
