% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-non-catalogue-data.R
\name{mapsheets_250K}
\alias{mapsheets_250K}
\title{NTS 250K Grid - Digital Baseline Mapping at 1:250,000 (NTS)}
\source{
https://open.canada.ca/data/en/dataset/055919c2-101e-4329-bfd7-1d0c333c0e62
}
\usage{
mapsheets_250K(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{mapsheets_250K} in the desired class
}
\description{
NTS 250K Grid - Digital Baseline Mapping at 1:250,000 (NTS)
}
\examples{
\dontrun{
my_layer <- mapsheets_250K()
my_layer_sp <- mapsheets_250K(class = 'sp')
}

}
