% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-non-catalogue-data.R
\name{mapsheets_50K}
\alias{mapsheets_50K}
\title{NTS 50K Grid - Digital Baseline Mapping at 1:50,000 (NTS)}
\source{
https://open.canada.ca/data/en/dataset/055919c2-101e-4329-bfd7-1d0c333c0e62
}
\usage{
mapsheets_50K()
}
\value{
The spatial layer of \code{mapsheets_50K} as an \code{sf} object.
}
\description{
NTS 50K Grid - Digital Baseline Mapping at 1:50,000 (NTS)
}
\examples{
\dontrun{
my_layer <- mapsheets_50K()
}

}
\seealso{
Other BC layers: 
\code{\link{airzones}()},
\code{\link{bc_bound}()},
\code{\link{bc_bound_hres}()},
\code{\link{bc_cities}()},
\code{\link{bec}()},
\code{\link{census_dissemination_area}()},
\code{\link{census_division}()},
\code{\link{census_economic}()},
\code{\link{census_metropolitan_area}()},
\code{\link{census_subdivision}()},
\code{\link{census_tract}()},
\code{\link{ecoprovinces}()},
\code{\link{ecoregions}()},
\code{\link{ecosections}()},
\code{\link{fsa}()},
\code{\link{gw_aquifers}()},
\code{\link{health_chsa}()},
\code{\link{health_ha}()},
\code{\link{health_hsda}()},
\code{\link{health_lha}()},
\code{\link{hydrozones}()},
\code{\link{mapsheets_250K}()},
\code{\link{municipalities}()},
\code{\link{nr_areas}()},
\code{\link{nr_districts}()},
\code{\link{nr_regions}()},
\code{\link{regional_districts}()},
\code{\link{tsa}()},
\code{\link{water_districts}()},
\code{\link{water_precincts}()},
\code{\link{watercourses_15M}()},
\code{\link{watercourses_5M}()},
\code{\link{wsc_drainages}()}
}
\concept{BC layers}
