% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{nr_districts}
\alias{nr_districts}
\title{British Columbia Natural Resource (NR) Districts}
\usage{
nr_districts(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{nr_districts} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
Type \code{?bcmaps.rdata::nr_districts} for details.
}
\examples{
\dontrun{
my_layer <- nr_districts()
my_layer_sp <- nr_districts(class = 'sp')
}

}
