% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{nr_regions}
\alias{nr_regions}
\title{British Columbia Natural Resource (NR) Regions}
\source{
\code{bcdata::bcdc_get_data(record = 'dfc492c0-69c5-4c20-a6de-2c9bc999301f', resource = 'ec636f64-9c5f-4704-8e66-2dd43032c9b5')}
}
\usage{
nr_regions(ask = interactive(), force = FALSE)
}
\arguments{
\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{nr_regions} as an \code{sf} object.
}
\description{
British Columbia Natural Resource (NR) Regions
}
\examples{
\dontrun{
my_layer <- nr_regions()
}

}
