% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{nr_regions}
\alias{nr_regions}
\title{British Columbia Natural Resource (NR) Regions}
\usage{
nr_regions(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{nr_regions} in the desired class
}
\description{
You must have the \code{bcmapsdata} package installed to use this function.
}
\details{
Type \code{?bcmapsdata::nr_regions} for details.
}
\examples{
\dontrun{
my_layer <- nr_regions()
my_layer_sp <- nr_regions(class = 'sp')
}

}
