% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{regional_districts}
\alias{regional_districts}
\title{British Columbia Ecoregions}
\usage{
regional_districts(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{regional_districts} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
See \link[bcmaps.rdata:regional_districts]{bcmaps.rdata::regional_districts} for details.
}
\examples{
\dontrun{
my_layer <- regional_districts()
my_layer_sp <- regional_districts(class = 'sp')
}

}
