% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_by_poly.R
\name{summarize_raster_list}
\alias{summarize_raster_list}
\title{Summarize a list of rasters into a list of numeric vectors}
\usage{
summarize_raster_list(raster_list, parallel = FALSE, cores = NULL, ...)
}
\arguments{
\item{raster_list}{list of rasters}

\item{parallel}{process in parallel? Default \code{FALSE}}

\item{cores}{number of cores if doing parallel. Default \code{NULL} uses half the number detected}

\item{...}{passed on to \code{doMC::registerDoMC}}
}
\value{
a list of numeric vectors
}
\description{
Summarize a list of rasters into a list of numeric vectors
}
