% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_by_poly.R
\name{summarize_raster_list}
\alias{summarize_raster_list}
\title{Summarize a list of rasters into a list of numeric vectors}
\usage{
summarize_raster_list(raster_list, parallel = FALSE)
}
\arguments{
\item{raster_list}{list of rasters}

\item{parallel}{process in parallel? Default \code{FALSE}. If \code{TRUE}, it is up to
the user to call \code{\link[future:plan]{future::plan()}} (or set \link[future:future.options]{options})
to specify what parallel strategy to use.}
}
\value{
a list of numeric vectors
}
\description{
Summarize a list of rasters into a list of numeric vectors
}
