% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_bc_albers}
\alias{transform_bc_albers}
\title{Transform a Spatial* object to BC Albers projection}
\usage{
transform_bc_albers(obj)
}
\arguments{
\item{obj}{The Spatial* or sf object to transform}
}
\value{
the Spatial* or sf object in BC Albers projection
}
\description{
Transform a Spatial* object to BC Albers projection
}
