% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_bc_albers}
\alias{transform_bc_albers}
\title{Transform a Spatial* object to BC Albers projection}
\usage{
transform_bc_albers(obj)
}
\arguments{
\item{obj}{The sf object to transform.}
}
\value{
the sf object in BC Albers projection
}
\description{
The \code{Spatial} method has been removed as of bcmaps 2.0.0. The \code{sf} method is here to stay.
}
