% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{tsa}
\alias{tsa}
\title{British Columbia Timber Supply Areas and TSA Blocks}
\format{An \code{sf} or \code{Spatial} polygons object with B.C.'s Timber Supply
Areas and TSA Blocks}
\source{
Original data from the
\href{https://catalogue.data.gov.bc.ca/dataset/8daa29da-d7f4-401c-83ae-d962e3a28980}{B.C. Data Catalogue},
under the
\href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open
Government Licence - British Columbia}.
}
\usage{
tsa(class = c("sf", "sp"), ...)
}
\arguments{
\item{class}{class of object to import; one of \code{"sf"} (default) or \code{"sp"}.}

\item{...}{arguments passed on to \link{get_big_data}}
}
\description{
The spatial representation for a Timber Supply Area or TSA Supply Block:
A Timber Supply Area is the primary unit for allowable annual cut (AAC)
determination. A TSA Supply Block is a designated area within the TSA
where the Ministry approves the allowable annual cuts.
}
\details{
Updated 2017-11-03
}
