% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm-convert.R
\name{utm_convert}
\alias{utm_convert}
\title{Convert a data.frame of UTM coordinates to an sf object with a single CRS}
\usage{
utm_convert(x, xcol, ycol, zone, crs = "EPSG:3005", xycols = TRUE)
}
\arguments{
\item{x}{data.frame containing UTM coordinates, with a zone column}

\item{xcol}{the name of the 'easting' columnm}

\item{ycol}{the name of the 'northing' column}

\item{zone}{the name of the 'zone' column, or a single numeric value if
the data are all in one UTM zone}

\item{crs}{target CRS. Default BC Albers (EPSG:3005)}

\item{xycols}{should the X and Y columns be appended}
}
\value{
sf object in the chosen CRS
}
\description{
This can operate on a data frame containing coordinates from multiple zones
with a column denoting the zone, or a single zone for the full dataset.
}
\details{
This function assumes that your data are in the Northern Hemisphere and in
NAD83
}
