% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_districts.R
\docType{data}
\name{water_districts}
\alias{water_districts}
\title{Polygons delineating British Columbia's Water Management Districts. Water Districts are
administrative areas established by the WATER REGULATIONS.}
\format{A spatialPolygonsDataFrame with polygons delineating Water Management Districts for British Columbia.}
\source{
Original data from
  \href{https://catalogue.data.gov.bc.ca/dataset/water-management-districts}{},
   under the
  \href{http://www2.gov.bc.ca/gov/content/data/open-data/open-government-license-bc}{Open
  Government License - British Columbia}.
}
\usage{
water_districts
}
\description{
Polygons delineating British Columbia's Water Management Districts. Water Districts are
administrative areas established by the WATER REGULATIONS.
}
\keyword{datasets}
