% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{water_districts}
\alias{water_districts}
\title{British Columbia's Water Management Districts}
\source{
\code{bcdata::bcdc_get_data(record = '92cb3ad8-9582-48a9-9e79-9a9d33601e50', resource = '07f9aa3f-0b66-4a49-919f-332d12bcd8f0')}
}
\usage{
water_districts(ask = interactive(), force = FALSE)
}
\arguments{
\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{water_districts} as an \code{sf} object.
}
\description{
British Columbia's Water Management Districts
}
\examples{
\dontrun{
my_layer <- water_districts()
}

}
