% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{water_districts}
\alias{water_districts}
\title{British Columbia's Water Management Districts}
\usage{
water_districts(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{water_districts} in the desired class
}
\description{
You must have the \code{bcmapsdata} package installed to use this function.
}
\details{
Type \code{?bcmapsdata::water_districts} for details.
}
\examples{
\dontrun{
my_layer <- water_districts()
my_layer_sp <- water_districts(class = 'sp')
}

}
