% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_precincts.R
\docType{data}
\name{water_precincts}
\alias{water_precincts}
\title{Polygons delineating British Columbia's Water Management Precincts. Water Precincts are
jurisdictional areas within a Water District.}
\format{A spatialPolygonsDataFrame with polygons delineating Water Management Precincts for British Columbia.}
\source{
Original data from
  \href{https://catalogue.data.gov.bc.ca/dataset/water-management-precincts}{},
   under the
  \href{http://www2.gov.bc.ca/gov/content/data/open-data/open-government-license-bc}{Open
  Government License - British Columbia}.
}
\usage{
water_precincts
}
\description{
Polygons delineating British Columbia's Water Management Precincts. Water Precincts are
jurisdictional areas within a Water District.
}
\keyword{datasets}
