% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{water_precincts}
\alias{water_precincts}
\title{British Columbia's Water Management Precincts}
\usage{
water_precincts(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{water_precincts} in the desired class
}
\description{
You must have the \code{bcmapsdata} package installed to use this function.
}
\details{
Type \code{?bcmapsdata::water_precincts} for details.
}
\examples{
\dontrun{
my_layer <- water_precincts()
my_layer_sp <- water_precincts(class = 'sp')
}

}
