% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{watercourses_15M}
\alias{watercourses_15M}
\title{British Columbia watercourses at 1:15M scale.}
\usage{
watercourses_15M(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer in the desired class
}
\description{
Watercourses at 1:15M scale for British Columbia sourced from Natural Resources Canada GeoGratis data.
}
\details{
An sf or Spatial object with watercourses at the 1:15M scale for British Columbia.

Original data from
\href{https://www.nrcan.gc.ca/earth-sciences/geography/topographic-information/free-data-geogratis/download-directory-documentation/17215}{},
under the
\href{http://open.canada.ca/en/open-government-licence-canada}{Open
Government License - Canada}.
}
