% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-non-catalogue-data.R
\name{watercourses_15M}
\alias{watercourses_15M}
\title{British Columbia watercourses at 1:15M scale}
\source{
https://ftp.maps.canada.ca/pub/nrcan_rncan/vector/canvec/fgdb/Hydro/canvec_15M_CA_Hydro_fgdb.zip
}
\usage{
watercourses_15M(class = "sf", ask = interactive(), force = FALSE)
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}

\item{ask}{Should the function ask the user before downloading the data to a cache? Defaults to the value of interactive().}

\item{force}{Should you force download the data?}
}
\value{
The spatial layer of \code{watercourses_15M} in the desired class
}
\description{
British Columbia watercourses at 1:15M scale
}
\examples{
\dontrun{
my_layer <- watercourses_15M()
my_layer_sp <- watercourses_15M(class = 'sp')
}

}
