% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{watercourses_15M}
\alias{watercourses_15M}
\title{British Columbia watercourses at 1:15M scale}
\usage{
watercourses_15M(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{watercourses_15M} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
Type \code{?bcmaps.rdata::watercourses_15M} for details.
}
\examples{
\dontrun{
my_layer <- watercourses_15M()
my_layer_sp <- watercourses_15M(class = 'sp')
}

}
