% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watercourses_5M.R
\docType{data}
\name{watercourses_5M}
\alias{watercourses_5M}
\title{Watercourses at 1:5M scale for British Columbia sourced from Natural Resources Canada GeoGratis data.}
\format{A spatialPolygonsDataFrame with watercourses at the 1:5M scale for British Columbia.}
\source{
Original data from
  \href{https://www.nrcan.gc.ca/earth-sciences/geography/topographic-information/free-data-geogratis/download-directory-documentation/17215}{},
   under the
  \href{http://open.canada.ca/en/open-government-licence-canada}{Open
  Government License - Canada}.
}
\usage{
watercourses_5M
}
\description{
Watercourses at 1:5M scale for British Columbia sourced from Natural Resources Canada GeoGratis data.
}
\keyword{datasets}

