% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{watercourses_5M}
\alias{watercourses_5M}
\title{British Columbia watercourses at 1:5M scale}
\usage{
watercourses_5M(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{watercourses_5M} in the desired class
}
\description{
You must have the \code{bcmapsdata} package installed to use this function.
}
\details{
Type \code{?bcmapsdata::watercourses_5M} for details.
}
\examples{
\dontrun{
my_layer <- watercourses_5M()
my_layer_sp <- watercourses_5M(class = 'sp')
}

}
