% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watersheds.R
\docType{data}
\name{watersheds}
\alias{watersheds}
\title{British Columbia Hydrometric Watershed Boundaries}
\format{A spatialPolygonsDataFrame with the boundaries of 457 watersheds in British Columbia.}
\source{
Original data from
  \href{http://catalogue.data.gov.bc.ca/dataset/hydrology-hydrometric-watershed-boundaries}{DataBC},
   under the
  \href{http://www.data.gov.bc.ca/local/dbc/docs/license/OGL-vbc2.0.pdf}{Open
  Government License - British Columbia}.
}
\usage{
watersheds
}
\description{
Watershed boundary delineated for Canada-BC hydrometric stations. 
Currently, watersheds were delineated using 1:50,000 scale boundaries in 1996, 
and many watersheds encompass entire drainages, instead of just the upstream watersheds.
}
\keyword{datasets}

