% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{wsc_drainages}
\alias{wsc_drainages}
\title{For use in showing the 4th level drainages (sub-sub-drainages) in the national Drainage Area Framework and for use in determining the Water Survey of Canada alphanumeric codes that are used in hydrometric station numbering.}
\usage{
wsc_drainages(class = "sf")
}
\arguments{
\item{class}{what class you want the object in? \code{"sf"} (default) or \code{"sp"}.}
}
\value{
The spatial layer of \code{wsc_drainages} in the desired class
}
\description{
You must have the \code{bcmaps.rdata} package installed to use this function.
}
\details{
See \link[bcmaps.rdata:wsc_drainages]{bcmaps.rdata::wsc_drainages} for details.
}
\examples{
\dontrun{
my_layer <- wsc_drainages()
my_layer_sp <- wsc_drainages(class = 'sp')
}

}
